+++ CHANGELOG.md
@@ -5,6 +5,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Adding herited classes that contains helper methods.
+
+### Changed
+
+- Moving the class based resources from nested modules to root module.
+- Moving LocalizedData of class based resources in .strings.psd1 files.
+Based on [stylesguidelines](https://dsccommunity.org/styleguidelines/localization/) of DscCommunity.
+
+### Removed
+
+- Removing dummy object
+
 ## [0.7.2] - 2020-09-29
 
 ### Changed